Base all transformations on the Transform object.
It has methods to generate appropriate CSS rules and hashed class names, and an API that can accomodate both existing functionality and new transformations better.

The external API now mainly consists of the Firmin.transform method, which takes an element to be transformed, and an object stating which operations should be performed, and their values.