Implement a common parsing and style building API for Transforms and Transitions.
Both types of object now have a parse method that accepts an animation description and returns an object with two properties: the result (either a Transform or a Transition), and the remainder (any unused properties of the description).

They also both implement similar build methods which take an object representing a CSS style and return a new style object with the appropriate transform and transition properties.